More spacing issues, still not quite resolved.
Attempt to solve it through a space-gobble phase, but (a) it slows it down
a little, and (b) it's not quite right yet. The easy way out would be to
add spaces before and after every new element, and then gobble it, but
that's both inefficient and inelegant (but it gets the spacing right...)

The tonality() function still doesn't work.

Added the Getchar() function from foo.vim, which takes care of the
left/right/etc. keys.