Ready to make some serious changes
Spacing, spacing, spacing.... Currently input is done by moving the cursor
pos. to the following whitespace, then back again to the beginning of the
previous note string. I believe this is what is creating problems some
times.

What I intend to do in the test branch, is to make the put_note_string()
procedure into a function, and to simplify the positioning operation at
the start of the function